requirements.txt
PyPDF2
sentence-transformers
faiss-cpu
transformers
streamlit
regex
numpy
pandas
scikit-learn